src/requirement.txt
coverage==4.5.1
Django==1.9
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
MySQL-python==1.2.5
olefile==0.44
Pillow==4.3.0
selenium==3.11.0
sqlparse==0.2.4
